+++ src/backend/joanie/tests/core/api/order/test_create.py
@@ -96,7 +96,6 @@ class OrderCreateApiTest(BaseAPITestCase):
             "course_code": course.code,
             "organization_id": str(organization.id),
             "product_id": str(product.id),
-            "has_consent_to_terms": True,
         }
         token = self.get_user_token("panoramix")
 
@@ -240,7 +239,6 @@ class OrderCreateApiTest(BaseAPITestCase):
             "enrollment_id": str(enrollment.id),
             "organization_id": str(organization.id),
             "product_id": str(product.id),
-            "has_consent_to_terms": True,
         }
         token = self.generate_token_from_user(enrollment.user)
 
@@ -411,7 +409,6 @@ class OrderCreateApiTest(BaseAPITestCase):
             "enrollment_id": str(enrollment.id),
             "organization_id": str(organization.id),
             "product_id": str(product.id),
-            "has_consent_to_terms": True,
         }
         token = self.get_user_token("panoramix")
 
@@ -450,7 +447,6 @@ class OrderCreateApiTest(BaseAPITestCase):
         data = {
             "course_code": course.code,
             "product_id": str(product.id),
-            "has_consent_to_terms": True,
         }
         token = self.get_user_token("panoramix")
 
@@ -480,7 +476,6 @@ class OrderCreateApiTest(BaseAPITestCase):
         data = {
             "course_code": course.code,
             "product_id": str(product.id),
-            "has_consent_to_terms": True,
         }
         token = self.get_user_token("panoramix")
 
@@ -542,7 +537,6 @@ class OrderCreateApiTest(BaseAPITestCase):
         data = {
             "course_code": course.code,
             "product_id": str(product.id),
-            "has_consent_to_terms": True,
         }
         token = self.get_user_token("panoramix")
 
@@ -727,7 +721,6 @@ class OrderCreateApiTest(BaseAPITestCase):
             "product_id": str(product.id),
             "id": uuid.uuid4(),
             "amount": 0.00,
-            "has_consent_to_terms": True,
         }
         token = self.get_user_token("panoramix")
 
@@ -867,7 +860,6 @@ class OrderCreateApiTest(BaseAPITestCase):
             "course_code": course.code,
             "organization_id": str(organization.id),
             "product_id": str(product.id),
-            "has_consent_to_terms": True,
         }
         token = self.get_user_token("panoramix")
 
@@ -916,7 +908,6 @@ class OrderCreateApiTest(BaseAPITestCase):
             "course_code": course.code,
             "organization_id": str(organization.id),
             "product_id": str(product.id),
-            "has_consent_to_terms": True,
         }
         token = self.get_user_token("panoramix")
 
@@ -967,7 +958,6 @@ class OrderCreateApiTest(BaseAPITestCase):
             response.json(),
             {
                 "product_id": ["This field is required."],
-                "has_consent_to_terms": ["This field is required."],
             },
         )
 
@@ -978,7 +968,6 @@ class OrderCreateApiTest(BaseAPITestCase):
             HTTP_AUTHORIZATION=f"Bearer {token}",
             data={
                 "product_id": str(product.id),
-                "has_consent_to_terms": True,
             },
         )
 
@@ -989,58 +978,6 @@ class OrderCreateApiTest(BaseAPITestCase):
             response.json(),
             {"__all__": ["Either the course or the enrollment field is required."]},
         )
-
-    def test_api_order_create_authenticated_product_with_contract_require_terms_consent(
-        self,
-    ):
-        """
-        The payload must contain has_consent_to_terms sets to True to create an order.
-        """
-        relation = factories.CourseProductRelationFactory()
-        token = self.get_user_token("panoramix")
-        billing_address = BillingAddressDictFactory()
-
-        data = {
-            "product_id": str(relation.product.id),
-            "course_code": relation.course.code,
-            "billing_address": billing_address,
-        }
-
-        # - `has_consent_to_terms` is required
-        response = self.client.post(
-            "/api/v1.0/orders/",
-            content_type="application/json",
-            HTTP_AUTHORIZATION=f"Bearer {token}",
-            data=data,
-        )
-        self.assertContains(
-            response,
-            '{"has_consent_to_terms":["This field is required."]}',
-            status_code=HTTPStatus.BAD_REQUEST,
-        )
-
-        # - `has_consent_to_terms` must be set to True
-        data.update({"has_consent_to_terms": False})
-        response = self.client.post(
-            "/api/v1.0/orders/",
-            content_type="application/json",
-            HTTP_AUTHORIZATION=f"Bearer {token}",
-            data=data,
-        )
-        self.assertContains(
-            response,
-            '{"has_consent_to_terms":["You must accept the terms and conditions to proceed."]}',
-            status_code=HTTPStatus.BAD_REQUEST,
-        )
-
-        data.update({"has_consent_to_terms": True})
-        response = self.client.post(
-            "/api/v1.0/orders/",
-            content_type="application/json",
-            HTTP_AUTHORIZATION=f"Bearer {token}",
-            data=data,
-        )
-        self.assertEqual(response.status_code, HTTPStatus.CREATED)
 
     def test_api_order_create_authenticated_product_course_unicity(self):
         """
@@ -1060,7 +997,6 @@ class OrderCreateApiTest(BaseAPITestCase):
             "product_id": str(product.id),
             "course_code": course.code,
             "organization_id": str(organization.id),
-            "has_consent_to_terms": True,
         }
 
         response = self.client.post(
@@ -1103,7 +1039,6 @@ class OrderCreateApiTest(BaseAPITestCase):
             "product_id": str(product.id),
             "course_code": course.code,
             "organization_id": str(organization.id),
-            "has_consent_to_terms": True,
         }
 
         response = self.client.post(
@@ -1139,7 +1074,6 @@ class OrderCreateApiTest(BaseAPITestCase):
             "organization_id": str(organization.id),
             "product_id": str(product.id),
             "billing_address": billing_address,
-            "has_consent_to_terms": True,
         }
 
         with self.assertNumQueries(60):
@@ -1287,7 +1221,6 @@ class OrderCreateApiTest(BaseAPITestCase):
             "order_group_id": str(order_group.id),
             "product_id": str(product.id),
             "billing_address": billing_address,
-            "has_consent_to_terms": True,
         }
         token = self.generate_token_from_user(user)
 
@@ -1337,7 +1270,6 @@ class OrderCreateApiTest(BaseAPITestCase):
             "order_group_id": str(order_group.id),
             "product_id": str(product.id),
             "billing_address": billing_address,
-            "has_consent_to_terms": True,
         }
         token = self.generate_token_from_user(user)
 
@@ -1368,7 +1300,6 @@ class OrderCreateApiTest(BaseAPITestCase):
             "course_code": course.code,
             "organization_id": str(organization.id),
             "product_id": str(product.id),
-            "has_consent_to_terms": True,
         }
         response = self.client.post(
             "/api/v1.0/orders/",
@@ -1395,7 +1326,6 @@ class OrderCreateApiTest(BaseAPITestCase):
             "course_code": course.code,
             "organization_id": str(organization.id),
             "product_id": str(product.id),
-            "has_consent_to_terms": True,
             "billing_address": billing_address,
             "credit_card_id": str(credit_card.id),
         }
@@ -1432,7 +1362,6 @@ class OrderCreateApiTest(BaseAPITestCase):
             "organization_id": str(relation.organizations.first().id),
             "product_id": str(product.id),
             "billing_address": billing_address,
-            "has_consent_to_terms": True,
         }
         token = self.generate_token_from_user(user)
 
@@ -1476,7 +1405,6 @@ class OrderCreateApiTest(BaseAPITestCase):
             "organization_id": str(organization.id),
             "product_id": str(relation.product.id),
             "billing_address": billing_address,
-            "has_consent_to_terms": True,
         }
 
         response = self.client.post(
@@ -1527,7 +1455,6 @@ class OrderCreateApiTest(BaseAPITestCase):
             "organization_id": str(relation.organizations.first().id),
             "product_id": str(product.id),
             "billing_address": billing_address,
-            "has_consent_to_terms": True,
         }
         token = self.generate_token_from_user(user)
 
@@ -1582,7 +1509,6 @@ class OrderCreateApiTest(BaseAPITestCase):
             "organization_id": str(relation.organizations.first().id),
             "product_id": str(product.id),
             "billing_address": billing_address,
-            "has_consent_to_terms": True,
         }
         token = self.generate_token_from_user(user)